backend/requirements.txt
# API Dependencies
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
httpx==0.26.0

# Core Dependencies
openai>=1.12.0
pinecone-client==3.0.0
python-dotenv==1.0.0
PyPDF2==3.0.1
scikit-learn>=1.0.2
numpy>=1.21.0
fuzzywuzzy
python-Levenshtein
python-multipart 